requirements.txt
PySocks==1.6.8
python-gnupg==0.4.2
stem==1.6.0
